split step by step into two pages
splits step by step into pages "download" and "installation and setup".
updates old assets to match up-to-date UI.
optimize video assets to be friendlier to load on the web
tweaks introduction slightly
removes tutorial on button making and execution, as that is moreso for another section, will decide in a bit what to do with it.